change int to size_t

+++ algs4Cpp/algs4Cpp/MergeX.h
@@ -14,12 +14,12 @@ namespace algs4Cpp {
 	private:
 	    static const int CUTOFF = 7;
 
-		static void merge(std::vector<Item> &src, std::vector<Item> &dst, int lo, int mid, int hi, const std::function<bool(Item, Item)> &less) {
+		static void merge(std::vector<Item> &src, std::vector<Item> &dst, size_t lo, size_t mid, size_t hi, const std::function<bool(Item, Item)> &less) {
 			assert(isSorted(src, lo, mid,less));
 			assert(isSorted(src, mid + 1, hi,less));
 
-			int i = lo, j = mid + 1;
-			for (int k = lo; k <= hi; ++k) {
+			size_t i = lo, j = mid + 1;
+			for (size_t k = lo; k <= hi; ++k) {
 				if (i > mid) dst[k] = src[j++];
 				else if (j > hi) dst[k] = src[i++];
 				else if (less(src[j],src[i])) dst[k] = src[j++];
@@ -29,13 +29,13 @@ namespace algs4Cpp {
 			assert(isSorted(dst, lo, hi,less));
 		}
 
-		static void sort(std::vector<Item>& src, std::vector<Item>& dst, int lo, int hi, const std::function<bool(Item, Item)> &less) {
+		static void sort(std::vector<Item>& src, std::vector<Item>& dst, size_t lo, size_t hi, const std::function<bool(Item, Item)> &less) {
 			if (hi <= lo + CUTOFF) {
 				insertionSort(dst, lo, hi,less);
 				return;
 			}
 
-			int mid = lo + (hi - lo) / 2;
+			size_t mid = lo + (hi - lo) / 2;
 			sort(dst, src, lo, mid,less);
 			sort(dst, src, mid + 1, hi,less);
 
@@ -53,18 +53,18 @@ namespace algs4Cpp {
 			return isSorted(a, 0, a.size() - 1,less);
 		}
 
-		static bool isSorted(const std::vector<Item> &a, int lo, int hi, const std::function<bool(Item, Item)> &less) {
-			for (int i = lo + 1; i <= hi; ++i) {
+		static bool isSorted(const std::vector<Item> &a, size_t lo, size_t hi, const std::function<bool(Item, Item)> &less) {
+			for (size_t i = lo + 1; i <= hi; ++i) {
 				if (less(a[i],a[i - 1])) return false;
 			}
 			return true;
 		}
 
 
-		static void insertionSort(std::vector<Item> &a, int lo, int hi, const std::function<bool(Item, Item)> &less) {
+		static void insertionSort(std::vector<Item> &a, size_t lo, size_t hi, const std::function<bool(Item, Item)> &less) {
 			using std::swap;
-			for (int i = lo; i <= hi; ++i) {
-				for (int j = i; j > lo && less(a[j],a[j - 1]); --j) {
+			for (size_t i = lo; i <= hi; ++i) {
+				for (size_t j = i; j > lo && less(a[j],a[j - 1]); --j) {
 					swap(a[j], a[j - 1]);
 				}
 			}
