+++ algs4Cpp/algs4Cpp/Vector.h
@@ -8,7 +8,7 @@
 namespace algs4Cpp {
 	class Vector {
 	private:
-		int d;
+		size_t d;
 		std::vector<double> data;
 
 	public:
@@ -23,18 +23,18 @@ namespace algs4Cpp {
 			}
 		}
 
-		int length() const {
+		size_t length() const {
 			return d;
 		}
 
-		int dimension() const {
+		size_t dimension() const {
 			return d;
 		}
 
 		double dot(const Vector &that) const {
 			check(that);
 			double sum = 0.0;
-			for (int i = 0; i != d;++i) {
+			for (size_t i = 0; i != d;++i) {
 				sum += (data[i] * that.data[i]);
 			}
 			return sum;
@@ -51,7 +51,7 @@ namespace algs4Cpp {
 		Vector operator+(const Vector &that) const {
 			check(that);
 			Vector c(d);
-			for (int i = 0; i != d; ++i) {
+			for (size_t i = 0; i != d; ++i) {
 				c.data[i] = data[i] + that.data[i];
 			}
 			return c;
@@ -60,7 +60,7 @@ namespace algs4Cpp {
 		Vector operator-(const Vector &that) const {
 			check(that);
 			Vector c(d);
-			for (int i = 0; i != d; ++i) {
+			for (size_t i = 0; i != d; ++i) {
 				c.data[i] = data[i] - that.data[i];
 			}
 			return c;
@@ -68,9 +68,8 @@ namespace algs4Cpp {
 
 		bool operator==(const Vector &that) const {
 			if (dimension() != that.dimension()) return false;
-			for (int i = 0; i != d; ++i) {
+			for (size_t i = 0; i != d; ++i) {
 				if (data[i] != that.data[i]) return false;
-				//if (almostUnequal(data[i], that.data[i])) return false;
 			}
 			return true;
 		}
@@ -79,7 +78,7 @@ namespace algs4Cpp {
 			return !(*this == that);
 		}
 
-		double cartesian(int i) const {
+		double cartesian(size_t i) const {
 			return data[i];
 		}
 
@@ -89,7 +88,7 @@ namespace algs4Cpp {
 
 		Vector scale(double alpha) const {
 			Vector c(d);
-			for (int i = 0; i != d; ++i) {
+			for (size_t i = 0; i != d; ++i) {
 				c.data[i] = alpha*data[i];
 			}
 			return c;
@@ -104,19 +103,6 @@ namespace algs4Cpp {
 		void check(const Vector &that) const {
 			if (d != that.d) throw std::invalid_argument("Dimensions don't agree");
 		}
-
-		//bool almostEqual(double a, double b) const {
-		//	if (std::fabs(a) < std::fabs(b)) {
-		//		return std::fabs(a - b) <= fabs(b)*std::numeric_limits<double>::epsilon();
-		//	}
-		//	else {
-		//		return std::fabs(a - b) <= fabs(a)*std::numeric_limits<double>::epsilon();
-		//	}
-		//}
-
-		//bool  almostUnequal(double a, double b) const {
-		//	return !almostEqual(a, b);
-		//}
 
 	};
 }
